Cleaning up, modified input functions.
Created a process_input(...) function to take in const char* arrays.
This setup makes unit testing a little bit easier.

Swapped atoi for strtol(...), changed ints to long ints, as strtol
returns an long int.  The program will be just fine if it operates on
long ints as opposed to ints.