More cleaning.
Added more consts to function parameters.

Rearranged functions in fizz_buzz.  The run() function contains the meat
of the program, and thus should be first.  process_input, and
check_input are our input verification functions, and last, there is a
print_usage for printing errors.  This is printed if we get the wrong
number of inputs, as well as if our input is formatted incorrectly.